worker: split the sandbox runner from the Tenki provider and fix Freestyle setup
