+++ src/orchestrator/worker/sandbox-utils.test.ts
@@ -1,136 +1,88 @@
 import assert from 'node:assert/strict';
 import { describe, test } from 'node:test';
 
+import type { SandboxExecResult } from '../../types.js';
 import {
   assertExecSucceeded,
-  execExitCode,
-  execString,
+  execStderr,
+  execStdout,
   normalizeRemotePath,
   remoteJoin,
   shellQuote,
-} from './tenki-utils.js';
+} from './sandbox-utils.js';
 
 describe('assertExecSucceeded', () => {
-  const okCases: Array<{ name: string; result: unknown }> = [
-    { name: 'When the exit code is zero then should pass', result: { exitCode: 0 } },
-    // A plain string carries no exit code, so there is nothing to fail on.
-    { name: 'When the result is a string then should pass', result: 'some output' },
-    { name: 'When no exit code is present then should pass', result: { stdout: 'x' } },
-  ];
-
-  for (const testCase of okCases) {
-    test(testCase.name, () => {
-      assert.doesNotThrow(() => assertExecSucceeded(testCase.result as never, 'step'));
-    });
-  }
-
-  const failureCases: Array<{ name: string; result: Record<string, unknown>; want: RegExp }> = [
+  const cases: Array<{ name: string; result: SandboxExecResult; want?: RegExp }> = [
+    { name: 'When the exit code is zero then should succeed', result: execResult(0) },
     {
       name: 'When the exit code is not zero then should return error with the label and code',
-      result: { exitCode: 2 },
+      result: execResult(2),
       want: /step failed with exit code 2$/,
     },
     {
       name: 'When stderr is present then should append it as the detail',
-      result: { exitCode: 1, stderr: 'boom' },
-      want: /step failed with exit code 1: boom/,
+      result: execResult(1, '', 'boom'),
+      want: /step failed with exit code 1: boom$/,
     },
     {
       // stdout is the fallback detail: some commands report the reason there.
       name: 'When only stdout is present then should append it as the detail',
-      result: { exitCode: 1, stdout: 'no such file' },
-      want: /exit code 1: no such file/,
+      result: execResult(1, 'no such file'),
+      want: /step failed with exit code 1: no such file$/,
     },
     {
       name: 'When the detail is blank then should omit the suffix',
-      result: { exitCode: 1, stderr: '   ' },
-      want: /exit code 1$/,
+      result: execResult(1, '', '   '),
+      want: /step failed with exit code 1$/,
+    },
+    {
+      name: 'When the detail is longer than the cap then should truncate it',
+      result: execResult(1, '', 'x'.repeat(900)),
+      want: /step failed with exit code 1: x{500}$/,
     },
   ];
 
-  for (const testCase of failureCases) {
+  for (const testCase of cases) {
     test(testCase.name, () => {
-      assert.throws(() => assertExecSucceeded(testCase.result as never, 'step'), testCase.want);
+      if (!testCase.want) {
+        assert.doesNotThrow(() => assertExecSucceeded(testCase.result, 'step'));
+        return;
+      }
+      assert.throws(() => assertExecSucceeded(testCase.result, 'step'), testCase.want);
     });
   }
-
-  test('When the detail is huge then should truncate it', () => {
-    assert.throws(
-      () => assertExecSucceeded({ exitCode: 1, stderr: 'x'.repeat(900) } as never, 'step'),
-      (error: Error) => error.message.length < 600,
-    );
-  });
 });
 
-describe('execString', () => {
-  const cases: Array<{ name: string; result: unknown; keys: string[]; want: string }> = [
+describe('execStdout', () => {
+  const cases: Array<{ name: string; result: SandboxExecResult; want: string }> = [
     {
-      name: 'When the key holds a string then should return it',
-      result: { stderr: 'e' },
-      keys: ['stderr'],
-      want: 'e',
-    },
-    {
-      name: 'When the key holds a buffer then should decode it',
-      result: { stdout: Buffer.from('from buffer') },
-      keys: ['stdout'],
-      want: 'from buffer',
-    },
-    {
-      name: 'When the key holds bytes then should decode them',
-      result: { stdout: new TextEncoder().encode('from bytes') },
-      keys: ['stdout'],
+      name: 'When stdout carries bytes then should decode them',
+      result: execResult(0, 'from bytes'),
       want: 'from bytes',
     },
-    {
-      name: 'When the first key is absent then should try the next',
-      result: { output: 'second' },
-      keys: ['stdout', 'output'],
-      want: 'second',
-    },
-    {
-      name: 'When no key matches then should return empty',
-      result: { other: 'x' },
-      keys: ['stdout'],
-      want: '',
-    },
-    {
-      // A bare string result is stdout by convention, so only those keys read it.
-      name: 'When the result is a string and stdout is asked then should return it',
-      result: 'raw',
-      keys: ['stdout'],
-      want: 'raw',
-    },
-    {
-      name: 'When the result is a string and stderr is asked then should return empty',
-      result: 'raw',
-      keys: ['stderr'],
-      want: '',
-    },
+    { name: 'When stdout is empty then should return empty', result: execResult(0), want: '' },
   ];
 
   for (const testCase of cases) {
     test(testCase.name, () => {
-      assert.equal(execString(testCase.result as never, testCase.keys), testCase.want);
+      assert.equal(execStdout(testCase.result), testCase.want);
     });
   }
 });
 
-describe('execExitCode', () => {
-  const cases: Array<{ name: string; result: unknown; want?: number }> = [
-    { name: 'When exitCode is present then should return it', result: { exitCode: 3 }, want: 3 },
-    { name: 'When only code is present then should return it', result: { code: 4 }, want: 4 },
-    { name: 'When only status is present then should return it', result: { status: 5 }, want: 5 },
+describe('execStderr', () => {
+  const cases: Array<{ name: string; result: SandboxExecResult; want: string }> = [
     {
-      name: 'When the value is not a number then should return undefined',
-      result: { exitCode: '0' },
+      name: 'When stderr carries bytes then should decode them',
+      result: execResult(1, '', 'boom'),
+      want: 'boom',
     },
-    { name: 'When the result is a string then should return undefined', result: 'output' },
+    { name: 'When stderr is empty then should return empty', result: execResult(0), want: '' },
   ];
 
   for (const testCase of cases) {
     test(testCase.name, () => {
-      assert.equal(execExitCode(testCase.result as never), testCase.want);
+      assert.equal(execStderr(testCase.result), testCase.want);
     });
   }
 });
@@ -238,3 +190,8 @@ describe('remoteJoin', () => {
     });
   }
 });
+
+function execResult(exitCode: number, stdout = '', stderr = ''): SandboxExecResult {
+  const encoder = new TextEncoder();
+  return { exitCode, stdout: encoder.encode(stdout), stderr: encoder.encode(stderr) };
+}
